Guess formula paths by repository content
Instead of using a manually stored regex for the formula path we now
guess the right path for formulae. Usually formulae are expected to
reside under Formula/, but some taps are flat and contain all formulae
in the root.

Additionally, this patch should reduce the amount of noise that has to
be filtered with the regex like READMEs etc. because the path spec
given to Git also limits to Ruby files.